recipes-extended/ml-sdk/files/ARROW_DEMOS/TFLite-armnn/img_classification_demo/requirements.txt
requests>=2.24.0
urllib3>=1.25.8
Pillow>=6.2.2
numpy>=1.18.1